+++ service-implementation/src/main/java/com/liaison/mailbox/service/dto/configuration/IngressFilesDTO.java
@@ -1,0 +1,38 @@
+/**
+ * Copyright Liaison Technologies, Inc. All rights reserved.
+ *
+ * This software is the confidential and proprietary information of
+ * Liaison Technologies, Inc. ("Confidential Information").  You shall 
+ * not disclose such Confidential Information and shall use it only in
+ * accordance with the terms of the license agreement you entered into
+ * with Liaison Technologies.
+ */
+
+package com.liaison.mailbox.service.dto.configuration;
+
+import java.util.Date;
+
+/**
+ * DTO for ingress file.
+ */
+public class IngressFilesDTO {
+
+    private String path;
+    private Date createdTime;
+
+    public String getPath() {
+        return path;
+    }
+
+    public void setPath(String path) {
+        path = path;
+    }
+
+    public Date getCreatedTime() {
+        return createdTime;
+    }
+
+    public void setCreatedTime(Date createdTime) {
+        this.createdTime = createdTime;
+    }
+}