+++ service-implementation/src/main/java/com/liaison/mailbox/service/rest/FileIngressResource.java
@@ -1,0 +1,84 @@
+/**
+ * Copyright Liaison Technologies, Inc. All rights reserved.
+ *
+ * This software is the confidential and proprietary information of
+ * Liaison Technologies, Inc. ("Confidential Information").  You shall 
+ * not disclose such Confidential Information and shall use it only in
+ * accordance with the terms of the license agreement you entered into
+ * with Liaison Technologies.
+ */
+
+package com.liaison.mailbox.service.rest;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+import com.liaison.commons.acl.annotation.AccessDescriptor;
+import com.liaison.commons.audit.AuditStatement;
+import com.liaison.commons.audit.AuditStatement.Status;
+import com.liaison.commons.audit.DefaultAuditStatement;
+import com.liaison.commons.audit.hipaa.HIPAAAdminSimplification201303;
+import com.liaison.commons.audit.pci.PCIV20Requirement;
+import com.liaison.framework.AppConfigurationResource;
+import com.liaison.mailbox.service.core.FilesIngressSevice;
+import com.wordnik.swagger.annotations.Api;
+import com.wordnik.swagger.annotations.ApiOperation;
+import com.wordnik.swagger.annotations.ApiResponse;
+import com.wordnik.swagger.annotations.ApiResponses;
+
+/**
+ * This is the gateway for file ingress service.
+ * 
+ */
+
+@AppConfigurationResource
+@Path("config/fileingress")
+@Api(value = "config/fileingress", description = "File Ingress services")
+public class FileIngressResource extends AuditedResource {
+
+    private static final Logger LOG = LogManager.getLogger(FileIngressResource.class);
+
+    /**
+     * REST method to get ingress files from sweeper location.
+     * 
+     * @return Response Object
+     */
+    @GET
+    @ApiOperation(value = "File Ingress", notes = "List of ingress files in sweeper location", position = 1, response = com.liaison.mailbox.service.dto.configuration.response.FileIngressResponceDTO.class)
+    @Produces(MediaType.APPLICATION_JSON)
+    @ApiResponses({ @ApiResponse(code = 500, message = "Unexpected Service failure.") })
+    @AccessDescriptor(skipFilter = true)
+    public Response listIngressFiles(@Context final HttpServletRequest request) {
+
+        // create the worker delegate to perform the business logic
+        AbstractResourceDelegate<Object> worker = new AbstractResourceDelegate<Object>() {
+            @Override
+            public Object call() {
+
+                LOG.debug("Entering into file ingress resource.");
+                FilesIngressSevice filesIngressSevice = new FilesIngressSevice();
+                return filesIngressSevice.listIngressFiles();
+            }
+        };
+        worker.actionLabel = "FileIngressResource.listIngressFiles()";
+        // hand the delegate to the framework for calling
+        return process(request, worker);
+    }
+
+    @Override
+    protected AuditStatement getInitialAuditStatement(String actionLabel) {
+        return new DefaultAuditStatement(Status.ATTEMPT, actionLabel, PCIV20Requirement.PCI10_2_5,
+                PCIV20Requirement.PCI10_2_2, HIPAAAdminSimplification201303.HIPAA_AS_C_164_308_5iiD,
+                HIPAAAdminSimplification201303.HIPAA_AS_C_164_312_a2iv,
+                HIPAAAdminSimplification201303.HIPAA_AS_C_164_312_c2d);
+    }
+
+}